+++ lib/urlwatch/filters.py
@@ -117,8 +117,8 @@ class FilterBase(object, metaclass=TrackSubClasses):
             filter_spec = [dict([filter_kind.split(':', 1)]) if ':' in filter_kind else filter_kind
                            for filter_kind in filter_spec.split(',')]
 
-            logger.warn('String-based filter definitions (%s) are deprecated, please convert to dict-style (see https://urlwatch.readthedocs.io/en/latest/deprecated.html):\n\n%s',
-                        old_filter_spec, yaml.dump(filter_spec, default_flow_style=False))
+            logger.warning('String-based filter definitions (%s) are deprecated, please convert to dict-style (see https://urlwatch.readthedocs.io/en/latest/deprecated.html):\n\n%s',
+                           old_filter_spec, yaml.dump(filter_spec, default_flow_style=False))
 
         if isinstance(filter_spec, list):
             for item in filter_spec:
@@ -222,7 +222,7 @@ class LegacyHooksPyFilter(FilterBase):
                 result = data
             return result
         except Exception as e:
-            logger.warn('Could not apply legacy hooks filter: %s', e)
+            logger.warning('Could not apply legacy hooks filter: %s', e)
             return data